Stabilise UI tests and nightly

+++ tools/assets/build.sh
@@ -19,14 +19,20 @@ function build {
     echo "Parameter is empty"
     exit 1
   fi
+  if [[ ! -d $1 ]]; then
+     echo $1 folder not found
+     exit 1
+  fi
 
   pushd $1
   if [[ -d "node_modules" ]]; then
     rm -rf node_modules
   fi
 
+  touch buildLock
   npm ci
   npm run build
+  rm buildLock
   popd
 }
